Exception handling - delete

+++ src/main/java/com/educandoweb/course/services/UserService.java
@@ -4,10 +4,12 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.educandoweb.course.entities.User;
 import com.educandoweb.course.repositories.UserRepository;
+import com.educandoweb.course.services.exceptions.DatabaseException;
 import com.educandoweb.course.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -22,15 +24,25 @@ public class UserService {
 	
 	public User findById(Long id) {
 		Optional<User> obj =  repository.findById(id);
-		return obj.orElseThrow(() -> new ResourceNotFoundException(id));
+		return obj.orElseThrow(() -> new ResourceNotFoundException(id)); // Pode-se usar o try-catch
 	}
 	
 	public User insert(User obj) {
 		return repository.save(obj);
 	}
 	
-	public void Delete(Long id) {
-		repository.deleteById(id);
+	public void delete(Long id) {
+		if(!repository.existsById(id)) {
+			throw new ResourceNotFoundException(id);
+		}
+		try {
+			
+			repository.deleteById(id);
+			
+		} catch (DataIntegrityViolationException e) {
+			throw new DatabaseException(e.getMessage());
+		}
+		
 	}
 	
 	public User update(Long id, User obj) {
